refactor(common/models): move primary trie-compilation code into common/models/templates
We'll likely want to dynamically build a Trie to represent user-specific entries made available by the active OS.  We'll then blend _that_ with the 'static' distributed model.
