+++ common/models/templates/src/index.ts
@@ -4,4 +4,8 @@ export {
 } from "./common.js";
 export { default as QuoteBehavior } from "./quote-behavior.js";
 export { getLastPreCaretToken, Token, Tokenization, tokenize, wordbreak } from "./tokenization.js";
+export {
+  Entry, InternalNode, Leaf, Node
+} from './trie.js';
+export { addUnsorted, TrieBuilder } from './trie-builder.js';
 export { default as TrieModel, TrieModelOptions } from "./trie-model.js";
